more flexible annotator interface
Adds annotate_span, and makes the parser
use this new method. This method takes the
full token sequence and the span i and j.
This means the annotator has access to context,
and the span indexes. Meanwhile annotate only has
access to the token subsequence. The new method
annotate_span paves the way for context-based annotations
and joint annotation preprocesses.